smallfix: remove rounding errors on pattern svg visualization

+++ packages/pattern/wrappers.py
@@ -107,7 +107,7 @@ class VisPattern(core.ParametrizedPattern):
             the lower-right vertex coordinate for the convenice of future offsetting.
         """
         panel = self.pattern['panels'][panel_name]
-        vertices = np.asarray(panel['vertices'], dtype=int)
+        vertices = np.asarray(panel['vertices'])
         vertices = self._verts_to_px_coords(vertices)
         # Shift vertices for visibility
         vertices = vertices + offset
